+++ source/Examples/Resources/DSC_CMForestDiscovery/ForestDiscovery_Disabled.ps1
@@ -1,0 +1,17 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to disable AD Forest Discovery for Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMForestDiscovery ExampleSettings
+        {
+            SiteCode = 'Lab'
+            Enabled  = $false
+        }
+    }
+}